x64 overhaul: native 64-bit ABI for core runtime, canonical decorations, 32-bit path removal
- Remove all 32-bit host/target code paths (CoreRuntimeApi, RuntimeBundleResolver,
  MemoryPE, CrashHandler, TargetABI x64-only, WIN32 defines, dynamic_call.asm).
- DBDLLCore.cpp: re-apply and extend pointer-width migrations:
  * encryption family (Decrypt/Encrypt/EncryptDecrypt/UpdateFilenameFromVirtualTable
    as LPSTR; EncryptDBPro/EncryptWorkshopDBPro/EncryptNewFile/EncryptAllFiles as
    const char*) matching canonical PEAD/PEBD export decorations
  * string-returning exports Cl$/GetDate$/GetTime$/InkeyS/GetDXVer$/GetTypePattern$
    converted YAKK -> YA_K_K (native 64-bit handles)
  * array handle family: ArrayInsertAtBottom/Top/Element, ArrayDeleteElement,
    ArrayIndex*, Next/PreviousArrayIndex, ArrayCount/IndexValid/GetType,
    PushToStack/PopFromStack/AddToQueue/RemoveFromQueue, EmptyArray,
    IsArraySingleDim -> DWORD_PTR parameters; ref-table entries treated as
    uintptr_t (8-byte) matching CreateArray/ExpandArray; overlapping shuffles
    switched memcpy -> memmove; fixed EmptyArray header reads (slots 10..13)
    and a pStoreRefs leak in ArrayInsertAtTop
- CoreRuntimeApi.h: CoreSlotFree/CoreArrayFree types (slot address vs handle value)
- Tests: canonical-only mocks; PE32+ fixture; Refs32BitCoreImage; TargetABI x64-only
- Strip all stage_trace diagnostics from CBasic2DC/CFileC/CGfxC (DBDLLCore clean)
- FreeUptoDisplay teardown frees by slot address (root-cause heap corruption fix)

Verified: 886/886 unit tests green; t6/t7/t8/t9 smoke apps APP_EXIT=0.

+++ DBProCompiler/DBPCompilerEXE/DarkEXE.cpp
@@ -189,8 +189,8 @@ bool RunProgram(HINSTANCE hInstance, LPSTR* pReturnError)
 	bResult=CEXE.Run(bResult);
 
 	// Report Any Runtime Errors
-	DWORD dwRTError=CEXE.m_dwRuntimeErrorDWORD;
-	DWORD dwRTErrorLine=CEXE.m_dwRuntimeErrorLineDWORD;
+	DWORD dwRTError=static_cast<DWORD>(CEXE.m_dwRuntimeErrorDWORD);
+	DWORD dwRTErrorLine=static_cast<DWORD>(CEXE.m_dwRuntimeErrorLineDWORD);
 	if(dwRTError>0)
 	{
 		// create report string and store
@@ -434,10 +434,10 @@ void DumpDebugReport (
 	else
 	{
 		// regular runtime error
-		sprintf_s ( pLineToReport, _MAX_PATH, "%d", CEXE.m_dwRuntimeErrorDWORD );
+		sprintf_s ( pLineToReport, _MAX_PATH, "%d", static_cast<DWORD>(CEXE.m_dwRuntimeErrorDWORD) );
 	}
 	WritePrivateProfileStringA ( "CEXE", "m_dwRuntimeErrorDWORD", pLineToReport, pReportFile );
-	sprintf_s ( pLineToReport, _MAX_PATH, "%d", CEXE.m_dwRuntimeErrorLineDWORD );
+	sprintf_s ( pLineToReport, _MAX_PATH, "%d", static_cast<DWORD>(CEXE.m_dwRuntimeErrorLineDWORD) );
 	WritePrivateProfileStringA ( "CEXE", "m_dwRuntimeErrorLineDWORD", pLineToReport, pReportFile );			
 	WritePrivateProfileStringA(
 		"RUNTIME",
