+++ Dark Basic Public Shared/Dark Basic Pro SDK/Shared/Q3BSP/CBSPFile.cpp
@@ -784,12 +784,12 @@ bool Load ( char* szFilename, char* szMap )
 	// Uses actual or virtual file..
 	char VirtualFilename[_MAX_PATH];
 	strcpy(VirtualFilename, szFilename);
-	g_pGlob->UpdateFilenameFromVirtualTable( (DWORD)VirtualFilename);
+	g_pGlob->UpdateFilenameFromVirtualTable( VirtualFilename);
 
 	// Decrypt and use media, re-encrypt
-	g_pGlob->Decrypt( (DWORD)VirtualFilename );
+	g_pGlob->Decrypt( VirtualFilename );
 	bool bRes = LoadCore ( VirtualFilename, szMap );
-	g_pGlob->Encrypt( (DWORD)VirtualFilename );
+	g_pGlob->Encrypt( VirtualFilename );
 
 	return bRes;
